requirements.txt
nextcord==2.4.0
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.1
aiomysql==0.2.0
sqlalchemy==2.0.23
setuptools>=65.5.1
